Move oneline to renderer/oneline
