+++ internal/renderer/oneline/constats.go
