+++ internal/renderer/oneline/oneline.go
