+++ internal/renderer/oneline/render.go
