+++ internal/renderer/oneline/render_emoji.go
