+++ internal/renderer/oneline/render_solar.go
