+++ main.go
@@ -14,8 +14,8 @@ import (
 	"github.com/chubin/wttr.in/internal/ip"
 	"github.com/chubin/wttr.in/internal/location"
 	"github.com/chubin/wttr.in/internal/logging"
-	"github.com/chubin/wttr.in/internal/oneline"
 	"github.com/chubin/wttr.in/internal/renderer"
+	"github.com/chubin/wttr.in/internal/renderer/oneline"
 	"github.com/chubin/wttr.in/internal/server"
 	"github.com/chubin/wttr.in/internal/spec"
 	"github.com/chubin/wttr.in/internal/uplink"
